refactor: rename masternode/sync to node/sync. It does nothing to masternodes
