+++ src/coinjoin/server.cpp
@@ -8,7 +8,7 @@
 #include <evo/deterministicmns.h>
 #include <masternode/meta.h>
 #include <masternode/node.h>
-#include <masternode/sync.h>
+#include <node/sync.h>
 #include <net.h>
 #include <netmessagemaker.h>
 #include <net_processing.h>
