+++ src/evo/mnauth.cpp
@@ -10,7 +10,7 @@
 #include <llmq/utils.h>
 #include <masternode/meta.h>
 #include <masternode/node.h>
-#include <masternode/sync.h>
+#include <node/sync.h>
 #include <net.h>
 #include <netmessagemaker.h>
 #include <util/time.h>
