+++ src/init.cpp
@@ -98,7 +98,7 @@
 #include <masternode/active/notificationinterface.h>
 #include <masternode/meta.h>
 #include <masternode/node.h>
-#include <masternode/sync.h>
+#include <node/sync.h>
 #include <masternode/utils.h>
 #include <messagesigner.h>
 #include <netfulfilledman.h>
