+++ src/instantsend/instantsend.cpp
@@ -8,7 +8,7 @@
 #include <chainparams.h>
 #include <consensus/validation.h>
 #include <instantsend/signing.h>
-#include <masternode/sync.h>
+#include <node/sync.h>
 #include <node/blockstorage.h>
 #include <spork.h>
 #include <stats/client.h>
