+++ src/llmq/quorums.cpp
@@ -18,7 +18,7 @@
 #include <evo/deterministicmns.h>
 #include <evo/evodb.h>
 #include <masternode/node.h>
-#include <masternode/sync.h>
+#include <node/sync.h>
 #include <net.h>
 #include <netmessagemaker.h>
 #include <util/irange.h>
