+++ src/node/sync.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <masternode/sync.h>
+#include <node/sync.h>
 
 #include <chainparams.h>
 #include <governance/governance.h>
